+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## [Unreleased]
 
-## [1.0.0] - 2017-06-20
+## [1.1.0] - 2021-01-14
+### Added
+- Wraps request in bulk
+
+### Deprecated
+- Removes the id of the document. Let Elasticsearch server define it.
+
+## [1.0.0] - 2021-01-14
 ### Added
 - Sends doc to ELK serve